make autoMaterializeAssetEvaluationIds BigInteger instead of Ints (#25648)
BK

> Insert changelog entry or delete this section.

## Summary & Motivation

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.

+++ js_modules/dagster-ui/packages/ui-core/src/assets/AutoMaterializePolicyPage/GetEvaluationsQuery.tsx
@@ -122,7 +122,7 @@ export const GET_EVALUATIONS_QUERY = gql`
 export const GET_EVALUATIONS_SPECIFIC_PARTITION_QUERY = gql`
   query GetEvaluationsSpecificPartitionQuery(
     $assetKey: AssetKeyInput!
-    $evaluationId: Int!
+    $evaluationId: BigInt!
     $partition: String!
   ) {
     assetConditionEvaluationForPartition(
