+++ python_modules/dagster-graphql/dagster_graphql/schema/instigation.py
@@ -250,7 +250,7 @@ class GrapheneInstigationTick(graphene.ObjectType):
     requestedAssetKeys = non_null_list(GrapheneAssetKey)
     requestedAssetMaterializationCount = graphene.NonNull(graphene.Int)
     requestedMaterializationsForAssets = non_null_list(GrapheneRequestedMaterializationsForAsset)
-    autoMaterializeAssetEvaluationId = graphene.Field(graphene.Int)
+    autoMaterializeAssetEvaluationId = graphene.Field(graphene.BigInt)
     instigationType = graphene.NonNull(GrapheneInstigationType)
 
     class Meta:
